requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
azure-core==1.36.0
azure-storage-blob==12.27.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
fastapi==0.124.0
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
idna==3.11
isodate==0.7.2
pillow==12.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.0
uvicorn==0.38.0
